+++ src/test/java/nextstep/ladder/domain/ladder/LadderLineTest.java
@@ -1,5 +1,7 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
 
+import nextstep.ladder.domain.ladder.LadderLine;
+import nextstep.ladder.util.RandomConnectStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
